wip - render a single jumbo texture
